Adding a current term name endpoint for freerooms
